tools/search (main lineage): tmsearch compiles again against gbx::record's f32 samples; --plane takes an axis (x|y|z=VALUE), either crossing direction
main's tmsearch did not build: 13 f32/f64 mismatches in refline.rs and
seedstate.rs against the record decoder's f32 sample fields. Fixed at the use
sites. And --plane was an x-plane; colon three (TMX 353522) finishes on a
z-oriented road. The axis rides as a trailing u32 on the ARM wire (absent = x).

NOTE: this tree's tools/search is the pre-tickhook lineage (lroundf clock, the
map-2 clock fit). The tick-hook fork oracle that found the 7.590 lives on the
devvm42752 mirror (agentcloud/route, agentcloud/colon-three) and still has to
be brought to main.

+++ tools/search/tmsearch/src/refline.rs
@@ -104,7 +104,7 @@ pub fn from_ghost(
 
     let d = gbx::record::decode_ghost(path).map_err(|e| format!("{}: {}", path, e))?;
     let rows: Vec<(i64, f64, f64, f64)> =
-        d.samples.iter().map(|s| (s.time_ms as i64, s.x, s.y, s.z)).collect();
+        d.samples.iter().map(|s| (s.time_ms as i64, s.x as f64, s.y as f64, s.z as f64)).collect();
     let line = RefLineData::from_samples(&rows, start_offset_ms, nticks)
         .map_err(|e| format!("{}: {}", path, e))?;
     Ok(FromGhost { line, engine_error_m: mean, kappa, samples: rows.len() })
